tradovate-executor/requirements.txt
websockets>=12.0
httpx>=0.27.0
numpy>=1.26.0
cryptography>=42.0
pydantic>=2.5
python-dateutil>=2.9
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
pywebview>=4.0
